refactored layout
