+++ components/RightDrawer/index.tsx
@@ -1,0 +1,5 @@
+const RightDrawer: React.FC = () => {
+  return <div>Hello world</div>;
+};
+
+export default RightDrawer;
